% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm2}
\alias{hm2}
\title{HydroMonitor ObservationWell Data}
\format{
An object of class \code{list} of length 2.
}
\usage{
hm2
}
\description{
Sample dataset read with hm_read_export_csv2()
}
\keyword{datasets}
