% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm3}
\alias{hm3}
\title{Heads read from Dino *_1.csv file.}
\format{
An object of class \code{list} of length 2.
}
\usage{
hm3
}
\description{
Sample dataset read from Dino *_1.csv file.
}
\keyword{datasets}
