% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm4}
\alias{hm4}
\title{Measured heads read from Dino *_1.csv files in specified folder.}
\format{
An object of class \code{list} of length 2.
}
\usage{
hm4
}
\description{
Sample dataset read from Dino *_1.csv files in specified folder.
}
\keyword{datasets}
