% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm5}
\alias{hm5}
\title{Measured heads read from Dino zip file.}
\format{
An object of class \code{list} of length 2.
}
\usage{
hm5
}
\description{
Sample dataset read with Measured heads read from Dino zip file.
}
\keyword{datasets}
