% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_create_shp}
\alias{hm_create_shp}
\title{Create a point shape file from HydroMonitor ObservationWell data object.}
\usage{
hm_create_shp(hm, filename, crs = "+init=epsg:28992")
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{hm_read_export_csv}}}

\item{filename}{(character)}

\item{crs}{Character or object of class 'CRS'. PROJ.4 description of the coordinate reference system.
Default CRS is "Amersfoort / RD new".}
}
\value{
Nothing is returned when writing a shapeﬁle.
}
\description{
Create a summarizing point shape file from the meta data part of HydroMonitor ObservationWell data object.
}
\details{
(\code{\link{hm_read_export_csv}}).

Fields in the attribute table of the resulting shapefile are:
\itemize{
\item NAME: Name of observationwell.
\item X: X-coordinate of the observationwell.
\item Y: Y-coordinate of the observationwell.
\item NFILTERS: Number of filters.
\item TOP: Level of the top of the highest filter.
\item BOT: Level of the bottom of the lowest filter.
\item MV: Surface level.
}
}
\examples{
hm <- hm1
filename <- file.path(path.expand("~"),"filename.shp")
hm_create_shp(hm, filename)
}
