% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_create_shp_file}
\alias{hm_create_shp_file}
\title{Create a shape file from HydroMonitor ObservationWell data object.}
\usage{
hm_create_shp_file(hm, filename)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}

\item{filename}{(character)}
}
\value{
Nothing is returned when writing a shapeﬁle.
}
\description{
Create a shape file from the meta data part of HydroMonitor ObservationWell data object
(\code{\link{hm_read_export_csv}}).
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm <- hm_read_export_csv( fname )
filename <- file.path(path.expand("~"),"filename.shp")
hm_create_shp_file(hm, filename)
}
