% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_filter_on_extent}
\alias{hm_filter_on_extent}
\title{Filter HydroMonitor ObservationWell data on extent.}
\usage{
hm_filter_on_extent(hm, e)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}

\item{e}{Extent object or a Raster* or Spatial* object}
}
\value{
Filtered HydroMonitor ObservationWell data
}
\description{
Filter HydroMonitor ObservationWell data on extent.
}
\examples{
hm <- hm1
xmin <- min(hm$xm$X)
xmax <- max(hm$xm$X)
ymin <- min(hm$xm$Y)
ymax <- max(hm$xm$Y)
dx <- (xmax-xmin)/10
dy <- (ymax-ymin)/10
e <- raster::extent(c(xmin+dx, xmax-dx, ymin+dy, ymax-dy))
hm_filtered <- hm_filter_on_extent(hm, e)
}
