% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_filter_on_poly}
\alias{hm_filter_on_poly}
\title{Filter HydroMonitor ObservationWell data with polygon shape.}
\usage{
hm_filter_on_poly(hm, p, crs = "+init=epsg:28992")
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{hm_read_export_csv}}}

\item{p}{Polygon shape}

\item{crs}{Character or object of class 'CRS'. PROJ.4 description of the coordinate reference system.
Default CRS is "Amersfoort / RD new".}
}
\value{
HydroMonitor ObservationWell data within polygon shape.
}
\description{
Filter HydroMonitor ObservationWell data with polygon shape.
}
\details{
Only polygon shapes of length=1 can be used.
}
\examples{
hm <- hm1
p <- polygn
hm_filtered_on_polygon <- hm_filter_on_poly(hm, p)
}
