% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_read_export_csv.R
\name{hm_filter_on_year}
\alias{hm_filter_on_year}
\title{Filter HydroMonitor ObservationWell data on year.}
\usage{
hm_filter_on_year(hm, minyear = 1900, maxyear = 3000)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}

\item{minyear}{Minimal year to read data from (integer)}

\item{maxyear}{Maximal year to read data from (integer)}
}
\value{
Filtered HydroMonitor ObservationWell data \code{\link{hm_read_export_csv}}
}
\description{
Filter HydroMonitor ObservationWell data on year.
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm <- hm_read_export_csv( fname )
hm_filtered <- hm_filter_on_year( hm, minyear=2000)
}
