% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_gxg_table}
\alias{hm_gxg_table}
\title{Calculate GxG table based on head observations in top filters of HydroMonitor ObservationWell data.}
\usage{
hm_gxg_table(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{hm_read_export_csv}}}
}
\value{
Tibble with parameters as specified in \code{link{hm_calc_gxg}}.
}
\description{
Calculate GxG table based on head observations in top filters of HydroMonitor ObservationWell data.
}
\examples{
hm <- hm1
gxg <- hm_gxg_table(hm)
}
\seealso{
\code{link{hm_calc_gxg}}
}
