% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_plot}
\alias{hm_plot}
\title{Plot HydroMonitor ObservationWell data and optionally save plots to specified folder.}
\usage{
hm_plot(hm, output_dir = NULL)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}

\item{output_dir}{folder name to write plot (character). If NULL, plots are not written do disk.}
}
\value{
tibble. Fields:
\itemize{
\item NAME Name of observationwell (character vector)
\item plots List of timeseries plots (ggplot object)
}
}
\description{
Create a list of timeseries plots of all HydroMonitor Observationwell data
(\code{\link{hm_read_export_csv}}).
}
\examples{
hm <- hm1
x <- hm_plot(hm)
x$NAME[1]
x$plots[[1]]
}
