% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_rbind}
\alias{hm_rbind}
\title{Merge HydroMonitor ObservationWell data objects.}
\usage{
hm_rbind(..., hm_list = NULL)
}
\arguments{
\item{...}{HydroMonitor ObservationWell data objects.}

\item{hm_list}{Optional list of HydroMonitor ObservationWell data objects.}
}
\value{
HydroMonitor ObservationWell data object.
}
\description{
Bind rows of two HydroMonitor ObservationWell data objects (\code{\link{hm_read_export_csv}}).
}
\details{
Double observations and double filters are removed from the HMOW data.
}
\examples{
hm <- hm_rbind(hm1, hm2)
}
