% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_read_dino.R
\name{hm_read_dino}
\alias{hm_read_dino}
\title{Read export Dino *_1.csv file with measured heads.}
\usage{
hm_read_dino(fname)
}
\arguments{
\item{fname}{Filename (csv file, character)}
}
\value{
List of 2:
\itemize{
\item xm Characteristics of the monitoring well (meta data).
\item xd Measured heads (data.frame).
}

Variables in data.frame xm:
\itemize{
\item NAME Name of observationwell.  (character)
\item FILTER Filter number (integer)
\item X x-coordinate of observationwell (numeric)
\item Y y-coordinate of observationwell (numeric)
\item TOP Level of the top of filter (m+REF, numeric)
\item BOT Level of the bottom of filter (m+REF,numeric)
\item MV Surface level (m+REF,numeric)
}

Variables in data.frame xd:
\itemize{
\item NAME Name of observationwell (character)
\item FILTER Filter number (integer)
\item DATE Date of observation (POSIXct)
\item HEAD Observed head (numeric)
}
}
\description{
Trailing letters of observation well are removed.
}
\examples{
fname <- system.file("extdata","B38E0108001_1.csv",package="menyanthes")
hm3 <- hm_read_dino( fname )
}
