% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_read_dino.R
\name{hm_read_dino_path}
\alias{hm_read_dino_path}
\title{Read all export Dino *_1.csv files with measured heads in specified folder.}
\usage{
hm_read_dino_path(path)
}
\arguments{
\item{path}{a character vector containing a single path name. Tilde expansion (see \code{\link{path.expand}}) is done.}
}
\description{
Read all export Dino *_1.csv files with measured heads in specified folder.
}
\examples{
path <- system.file("extdata","Grondwaterstanden_Put",package="menyanthes")
hm4 <- hm_read_dino_path( path )
}
