% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_read_export_csv.R
\name{hm_read_export_csv}
\alias{hm_read_export_csv}
\title{Read export HydroMonitor file with ObservationWell data.}
\usage{
hm_read_export_csv(fname)
}
\arguments{
\item{fname}{Filename of export of HydroMonitor file with ObservationWell data (csv file, character)}
}
\value{
List with two elements:
xm: Characteristics of monitoring well (meta data in data frame)
xd: Measured heads (data frame)
}
\description{
Read export HydroMonitor file with ObservationWell data.
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm1 <- hm_read_export_csv( fname )
}
