% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_read_export_csv.R
\name{hm_read_export_csv2}
\alias{hm_read_export_csv2}
\title{Read export HydroMonitor file with ObservationWell data.}
\usage{
hm_read_export_csv2(fname)
}
\arguments{
\item{fname}{Filename of export of HydroMonitor file with ObservationWell data (csv file, character)}
}
\value{
List with two elements:
xm: Characteristics of monitoring well (meta data in data frame)
xd: Measured heads (data frame)
}
\description{
Read export HydroMonitor file with ObservationWell data.
}
\examples{
fname <- system.file("extdata","Topsoil1.csv",package="menyanthes")
hm2 <- hm_read_export_csv2( fname )
}
