% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_rm_fltrs_with_no_obs}
\alias{hm_rm_fltrs_with_no_obs}
\title{Remove meta data of filters if there are no observations available.}
\usage{
hm_rm_fltrs_with_no_obs(hm)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}
}
\value{
HydroMonitor ObservationWell data where in meta data filters with no observations are removed.
}
\description{
Remove meta data of filters if there are no observations available.
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm <- hm_read_export_csv( fname )
hm_clean <- hm_rm_fltrs_with_no_obs( hm )
}
