% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menyanthes.R
\docType{package}
\name{menyanthes}
\alias{menyanthes}
\title{menyanthes: A package read / write / manipulate \href{https://www.kwrwater.nl/tools-producten/hydromonitor/}{HydroMonitor} Observationwell data.}
\description{
This package exports the following functions:
}
\details{
\itemize{
\item \code{\link{hm_read_export_csv}}
\item \code{\link{hm_read_export_csv2}}
\item \code{\link{hm_read_dino}}
\item \code{\link{hm_read_dino_path}}
\item \code{\link{hm_filter_on_year}}
\item \code{\link{hm_filter_on_extent}}
\item \code{\link{hm_rm_fltrs_with_no_obs}}
\item \code{\link{hm_rbind}}
\item \code{\link{hm_rm_dble_fltrs}}
\item \code{\link{hm_rm_dble_obs}}
\item \code{\link{hm_plot}}
\item \code{\link{hm_calc_gxg}}
\item \code{\link{hm_create_shp}}
\item \code{\link{nr_obs_ratio}}
}

This package exports the following sample data sets:
hm1, ..., hm4: HydroMonitor ObservationWell Data.
}
