% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygn}
\alias{polygn}
\title{Polygon shape to be used in examples.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 1 columns.
}
\usage{
polygn
}
\description{
Polygon shape to be used in examples, for example the function hm_filter_on_poly.
}
\keyword{datasets}
