% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lees_menyanthes_csv.R
\name{read_menyanthes_csv}
\alias{read_menyanthes_csv}
\title{Read Menyanthes csv data and calculate GxG's of specified years}
\usage{
read_menyanthes_csv(fname, minyear = 1900, maxyear = 3000)
}
\arguments{
\item{fname}{Filename of Menyanthes csv file (character)}

\item{minyear}{Minimal year to read data from (integer)}

\item{maxyear}{Maximal year to read data from (integer)}
}
\value{
List with two elements:
x: characteristics of monitoring well and GxG's of specified years (data frame)
xv: Measured heads (data frame)
}
\description{
Read Menyanthes csv data and calculate GxG's of specified years
}
