% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_read_export_csv.R
\name{hm_calc_gxg}
\alias{hm_calc_gxg}
\title{Calculate GxG's of HydroMonitor ObservationWell data}
\usage{
hm_calc_gxg(hm)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}
}
\value{
xm: Characteristics of monitoring well (meta data in data frame) with GxG information added.
}
\description{
Calculate GxG's of HydroMonitor ObservationWell data
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm <- hm_read_export_csv( fname )
gxg <- hm_calc_gxg(hm)
}
