% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_filter_on_poly}
\alias{hm_filter_on_poly}
\title{Filter HydroMonitor ObservationWell data with polygon shape.}
\usage{
hm_filter_on_poly(hm, p)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}

\item{p}{Polygon shape}
}
\value{
HydroMonitor ObservationWell data within polygon shape.
}
\description{
Filter HydroMonitor ObservationWell data with polygon shape.
}
\details{
Only polygon shapes of length=1 can be used.
}
\examples{
hm <- hm1
p <- polygn
hm_filtered_on_polygon <- hm_filter_on_poly(hm, p)
}
