% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_rbind}
\alias{hm_rbind}
\title{Merge two HydroMonitor ObservationWell data objects.}
\usage{
hm_rbind(hm1, hm2)
}
\arguments{
\item{hm1}{First HydroMonitor ObservationWell data object to be merged.}

\item{hm2}{Second HydroMonitor ObservationWell data object to be merged.}
}
\value{
HydroMonitor ObservationWell data object.
}
\description{
Bind rows of two HydroMonitor ObservationWell data objects (\code{\link{hm_read_export_csv}}).
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm1 <- hm_read_export_csv( fname )

}
