% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_rbind}
\alias{hm_rbind}
\title{Merge HydroMonitor ObservationWell data objects.}
\usage{
hm_rbind(hm_list)
}
\arguments{
\item{hm_list}{List of HydroMonitor ObservationWell (HMOW) data objects.}
}
\value{
HydroMonitor ObservationWell data object.
}
\description{
Merge HydroMonitor ObservationWell data objects (\code{\link{hm_read_export_csv}}).
}
\details{
Double observations and double filters are removed from the HMOW data.
}
\examples{
hm <- hm_rbind(list(hm1, hm2))
}
