% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_rm_dble_fltrs}
\alias{hm_rm_dble_fltrs}
\title{Remove double filter information in meta data part of HydroMonitor ObservationWell data}
\usage{
hm_rm_dble_fltrs(hm)
}
\arguments{
\item{hm}{HydroMonitor ObservationWell data as read by \code{\link{hm_read_export_csv}}}
}
\value{
HydroMonitor ObservationWell data with double filter information remove from meta data part of
HydroMonitor ObservationWell data \code{\link{hm_read_export_csv}}
}
\description{
Remove double filter information in meta data part of HydroMonitor ObservationWell data
}
\examples{
fname <- system.file("extdata","export_data_menyanthes.csv",package="menyanthes")
hm <- hm_read_export_csv( fname )
hm_clean <- hm_rm_dble_fltrs( hm )
}
