% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilitys.R
\name{hm_rm_dble_obs}
\alias{hm_rm_dble_obs}
\title{Remove double observations in HydroMonitor ObservationWell data.}
\usage{
hm_rm_dble_obs(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{hm_read_export_csv}}}
}
\value{
HydroMonitor ObservationWell data with double records removed.
}
\description{
Remove double observations in HydroMonitor ObservationWell data.
}
\examples{
hm <- hm1
hm_clean <- hm_rm_dble_obs( hm )
}
